[CORE] Backport https://github.com/bitcoin/bitcoin/commit/381a46e38fa9262c868e5fee9ed0d1f71af4059e
