+++ src/coincontrol.h
@@ -7,6 +7,7 @@
 #ifndef BITCOIN_COINCONTROL_H
 #define BITCOIN_COINCONTROL_H
 
+#include "policy/feerate.h"
 #include "primitives/transaction.h"
 #include "script/standard.h"
