+++ src/init.cpp
@@ -29,6 +29,7 @@
 #include "masternodeman.h"
 #include "miner.h"
 #include "net.h"
+#include "policy/feerate.h"
 #include "policy/policy.h"
 #include "rpc/server.h"
 #include "script/standard.h"
