+++ src/net.h
@@ -13,6 +13,7 @@
 #include "limitedmap.h"
 #include "mruset.h"
 #include "netbase.h"
+#include "policy/feerate.h"
 #include "protocol.h"
 #include "random.h"
 #include "streams.h"
