+++ src/rpcwallet.cpp
@@ -10,6 +10,8 @@
 #include "core_io.h"
 #include "init.h"
 #include "net.h"
+#include "policy/feerate.h"
+#include "policy/policy.h"
 #include "netbase.h"
 #include "rpc/server.h"
 #include "timedata.h"
