+++ src/txmempool.h
@@ -11,6 +11,7 @@
 
 #include "amount.h"
 #include "coins.h"
+#include "policy/feerate.h"
 #include "primitives/transaction.h"
 #include "sync.h"
